# docs/swagger.yaml
basePath: /v1
definitions:
...
  clients.SalesforceConsentimento:
    properties:
      acao:
        description: 'legacy: optin | optout'
        type: string
      canal:
        type: string
      categoria:
        type: string
      data:
        type: string
      finalidade:
        type: string
      motivo:
        description: required when status=OUT / acao=optout
        type: string
      status:
        description: IN | OUT
        type: string
    type: object
...
  clients.SalesforceEndereco:
    properties:
      bairro:
        type: string
      cep:
        type: string
      cidade:
        type: string
      complemento:
        type: string
      estado:
        type: string
      logradouro:
        type: string
      pais:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.SalesforceWebhookDados:
    properties:
      accountId:
        type: string
      canalOrigem:
        type: string
      canalUltimaModificacao:
        type: string
      cidade:
        type: string
      complemento:
        type: string
      consentimento:
        items:
          $ref: '#/definitions/clients.SalesforceConsentimento'
        type: array
      dataNascimento:
        type: string
      deficiencia:
        type: string
      email:
        type: string
      endereco:
        $ref: '#/definitions/clients.SalesforceEndereco'
      escolaridade:
        type: string
      genero:
        type: string
      idioma:
        items:
          type: string
        type: array
      isTourist:
        type: boolean
      nacionalidade:
        type: string
      nome:
        type: string
      nomeExibicao:
        type: string
      nomeSocial:
        type: string
      passaporte:
        type: string
      raca:
        type: string
      rendaFamiliar:
        type: string
      telefone1:
        type: string
      telefone2:
        type: string
      telefone3:
        type: string
      telefoneInternacional:
        type: string
      telefonePrincipal:
        type: string
      tipoTelefone1:
        type: string
      tipoTelefone2:
        type: string
      tipoTelefone3:
        type: string
    type: object
  handlers.SalesforceWebhookRequestSwagger:
    properties:
      cpf:
        example: "52998224725"
        type: string
      dados:
        $ref: '#/definitions/handlers.SalesforceWebhookDados'
      evento:
        enum:
        - atualizacao
        - anonimizacao
        example: atualizacao
        type: string
      updatedAt:
        example: "2026-08-31T17:33:49Z"
        type: string
    type: object
  handlers.SalesforceWebhookResponse:
    properties:
      cpf:
        type: string
      evento:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: suporte@rio.rj.gov.br
    name: Suporte RMI
    url: http://www.rio.rj.gov.br
  description: API para gerenciamento de dados de cidadãos do Rio de Janeiro, incluindo
    autodeclaração de informações e verificação de contato.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API RMI
  version: "1.0"
paths:
...
  /webhooks/salesforce/cidadao:
    post:
      consumes:
      - application/json
      description: 'Webhook inbound: CPF + delta em dados (mesmo shape do GET Person
        Account). JWT Keycloak com azp em SALESFORCE_WEBHOOK_CLIENTS. evento=atualizacao
        (default) ou anonimizacao; updatedAt opcional. Campo ausente=não alterar;
        null=limpar; string vazia=valor vazio. Resposta 202 com CPF mascarado. Enfileira
        salesforce_sync sem GET de volta ao SF.'
      parameters:
      - description: CPF, evento e delta de campos alterados
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SalesforceWebhookRequestSwagger'
      produces:
      - application/json
      responses:
        "202":
          description: Job enfileirado
          schema:
            $ref: '#/definitions/handlers.SalesforceWebhookResponse'
        "400":
          description: CPF inválido, dados ausentes ou evento inválido
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: JWT inválido ou ausente
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: azp do JWT não está em SALESFORCE_WEBHOOK_CLIENTS
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Integração Salesforce não configurada
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Webhook Salesforce cidadão
      tags:
      - salesforce
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Tipo: Bearer token. Exemplo: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...